requirements.txt
paho-mqtt <= 1.7
pyyaml == 6.0